[story/VSPC-202] added IVSMedia and VSMedia mapped super class
