+++ vspace/src/main/java/edu/asu/diging/vspace/core/model/IVSImage.java
@@ -2,7 +2,7 @@ package edu.asu.diging.vspace.core.model;
 
 import java.util.List;
 
-public interface IVSImage extends IVSFile {
+public interface IVSImage extends IVSMedia {
 
 	String getFilename();
