+++ vspace/src/main/java/edu/asu/diging/vspace/core/model/impl/VSFile.java
@@ -14,18 +14,13 @@ import edu.asu.diging.vspace.core.model.IVSFile;
 
 @Entity
 
-public class VSFile extends VSpaceElement implements IVSFile {
+public class VSFile extends VSMedia implements IVSFile {
 
     @Id
     @GeneratedValue(generator = "file_id_generator")
     @GenericGenerator(name = "file_id_generator", parameters = @Parameter(name = "prefix", value = "FILE"), strategy = "edu.asu.diging.vspace.core.data.IdGenerator")
     private String id;
 
-    @Lob
-    private String filename;
-    @Lob
-    private String parentPath;
-    private String fileType;   
     private String originalFileName;
     private String fileDescription;
 
@@ -37,36 +32,6 @@ public class VSFile extends VSpaceElement implements IVSFile {
     @Override
     public void setId(String id) {
         this.id = id;
-    }
-    
-    @Override
-    public String getFilename() {
-        return filename;
-    }
-    
-    @Override
-    public void setFilename(String filename) {
-        this.filename = filename;
-    }
-    
-    @Override
-    public String getParentPath() {
-        return parentPath;
-    }
-    
-    @Override
-    public void setParentPath(String parentPath) {
-        this.parentPath = parentPath;
-    }
-
-    @Override
-    public String getFileType() {
-        return fileType;
-    }
-
-    @Override
-    public void setFileType(String fileType) {
-        this.fileType = fileType;
     }
     
     @Override
